+++ extension.js
@@ -1,5 +1,5 @@
 /* DownFall – Gnome Shell Extension
- * Copyright (C) 2019-2024 Benjamin S Osenbach
+ * Copyright (C) 2019-2026 Benjamin S Osenbach
  *
  * Inspired by Let It Snow (https://github.com/offlineric/gsnow).
  *
@@ -24,9 +24,11 @@ import Gio from 'gi://Gio';
 import GObject from 'gi://GObject';
 import GLib from 'gi://GLib';
 import Clutter from 'gi://Clutter';
+import Cogl from 'gi://Cogl';
 
 import {Extension} from 'resource:///org/gnome/shell/extensions/extension.js'
 import * as Main from 'resource:///org/gnome/shell/ui/main.js'
+import * as PopupMenu from 'resource:///org/gnome/shell/ui/popupMenu.js';
 import * as QuickSettings from 'resource:///org/gnome/shell/ui/quickSettings.js';
 const QuickSettingsMenu = Main.panel.statusArea.quickSettings;
 
@@ -47,11 +49,11 @@ var FallItem = GObject.registerClass({
 
     change(text, fontstring, color, shadow) {
       //don't style on each iteration of fall()
-      this.set_style(`color: ${color}; ${shadow}`);
-
-      this.get_clutter_text().set_font_name(fontstring);
-      
       this.set_text(text);
+      this.get_clutter_text().set_color(color);
+      this.get_clutter_text().set_font_name(fontstring);
+      //this.set_style(`color: ${color.to_string().substring(0,7)}`);
+      //this.set_style(`${shadow}`);
     }
     
     fall() {
@@ -97,14 +99,14 @@ var FallItem = GObject.registerClass({
 	}
       }
       
-      this.ease({
+      this.easeAsync({
       	x : this.endX,
       	y : this.endY,
       	duration : time,
       	mode : this.fim.ANIMATIONMODE,
       	rotation_angle_z : rotation,
       	onComplete : () => {this.finish()}
-      });
+      }).catch((err) => {});
       
     }
 
@@ -185,7 +187,8 @@ var FallItem = GObject.registerClass({
 	matritem.hide();
 
 	//show a new Matrix trail character every `time/n` milliseconds
-	this.matAddID = GLib.timeout_add(GLib.PRIORITY_LOW, time*(1+i)/n, () => {
+	//stop making new matritems after 1st fall
+	this.matAddID = GLib.timeout_add_once(GLib.PRIORITY_LOW, time*(1+i)/n, () => {
 		//set the matritem at the current FallItem position
 		let pos = this.get_position();
       		matritem.set_position(pos[0], pos[1]);
@@ -201,8 +204,6 @@ var FallItem = GObject.registerClass({
 			matritem.show();
 			return GLib.SOURCE_CONTINUE; //stopped on 'destroy' signal
 		});
-		
-		return GLib.SOURCE_REMOVE; //stop making new matritems after 1st fall
 	});
 
 	this.changeID = GLib.timeout_add(GLib.PRIORITY_LOW, time, () => {
@@ -229,25 +230,26 @@ var FallItem = GObject.registerClass({
     	  
     	  this.fim.pane3d.add_child(flare);
     	  flare.set_position(this.endX, this.endY);
-    	  flare.set_style(`color:${this.fim.FLRCOLOR}; ${this.fim.FLRSHADOW}`);
       	  flare.get_clutter_text().set_font_name(this.fim.FLRFONT);
+      	  flare.get_clutter_text().set_color(this.fim.FLRCOLOR);
+    	  flare.set_style(`${this.fim.FLRSHADOW}`);
     	  flare.set_text( this.fim.FLRDISP[ GLib.random_int_range(0, this.fim.FLRDISP.length) ] );
     	  
-    	  flare.ease({
+    	  flare.easeAsync({
     	    x : this.endX + Math.cos(angle)*speed/100 * this.monitor.width,
     	    y : this.endY + Math.sin(angle)*speed/100 * this.monitor.height + 1,
     	    duration : 2000,
 	    opacity: 0,
     	    mode : Clutter.AnimationMode.EASE_OUT_EXPO,
     	    onComplete : () => {flare.destroy()}
-    	  });
+	  }).catch((err) => {});
     	  
     	}
       }
       
       //reset the FallItem after a delay (reduces CPU load)    
-      this.idleID = GLib.idle_add(GLib.PRIORITY_LOW,
-      			() => {this.fall(); return GLib.SOURCE_REMOVE});
+      this.idleID = GLib.idle_add_once(GLib.PRIORITY_LOW,
+      			() => {this.fall()});
     }
 
     stopTimers() {
@@ -282,6 +284,11 @@ const FIM = GObject.registerClass({
 	this.RUNNING = false;
 	this.MATRIXTRAILSON = false;
 
+    	this.FALLCOLOR = Cogl.Color.init_from_hsl(0, 1, 0.5); //works, but white
+    	//this.FALLCOLOR = new Cogl.Color({ red: 255, green: 0, blue: 0, alpha: 255 }); //works, but white
+    	this.MATCOLOR = new Cogl.Color({red: 0, green: 0, blue: 0}); //black
+    	this.FLRCOLOR = new Cogl.Color({red: 0, green: 0, blue: 0}); //black
+
     	this.settings.connect('changed', this.settingsChanged.bind(this));
     }
 
@@ -292,7 +299,7 @@ const FIM = GObject.registerClass({
 	this.ANIMATIONMODE = this.settings.get_int('clutteranimmode')+1;
 
     	this.FALLITEMS = this.settings.get_strv("falltext");
-    	this.FALLCOLOR = this.settings.get_string('textcolor');
+	//this.changeColor(this.settings.get_string('textcolor'), this.FALLCOLOR);
 
     	this.FALLFONT = this.settings.get_string('textfont');
 
@@ -319,7 +326,7 @@ const FIM = GObject.registerClass({
     	this.MATRIXTRAILS = this.settings.get_boolean('matrixtrails');
     	if (this.MATRIXTRAILS) {
 	  this.MATDISP = this.settings.get_strv("matdisplay");
-    	  this.MATCOLOR = this.settings.get_string('matcolor');
+	  this.changeColor(this.settings.get_string('matcolor'), this.MATCOLOR);
 	  this.MATFONT = this.settings.get_string('matfont');
 	  if (this.settings.get_boolean('matshad')) {
 	    switch (this.settings.get_int('matshadtype')) {
@@ -339,7 +346,7 @@ const FIM = GObject.registerClass({
     	this.FIREWORKS = this.settings.get_boolean('fireworks');
     	if (this.FIREWORKS) {
     	  this.FLRDISP = this.settings.get_strv("flrdisplay");
-    	  this.FLRCOLOR = this.settings.get_string('flrcolor');
+	  this.changeColor(this.settings.get_string('flrcolor'), this.FLRCOLOR);
     	  this.FLRFONT = this.settings.get_string('flrfont');
 	  if (this.settings.get_boolean('flrshad')) {
 	    switch (this.settings.get_int('flrshadtype')) {
@@ -428,10 +435,25 @@ const FIM = GObject.registerClass({
 	this.MATRIXTRAILSON = false;
     }
 
+    changeColor(colorstring, myCoglColor) {
+    	let tc = new Cogl.Color({red: 0, green: 0, blue: 0});
+    	let temp = tc.from_string(colorstring);
+
+	// default to black if parsing fails
+	if (!ok) {
+		myCoglColor.init_from_4f(0, 0, 0, 1);
+	} else {
+		myCoglColor.init_from_4f(temp[1].red / 255,
+					temp[1].green / 255,
+					temp[1].blue / 255,
+					temp[1].alpha);
+	}
+    }
+
 });
 
 const FeatureToggle = GObject.registerClass(
-class FeatureToggle extends QuickSettings.QuickToggle {
+class FeatureToggle extends QuickSettings.QuickMenuToggle {
     _init(extensionObject, fim) {
         super._init({
 	    title: _('DownFall'),
@@ -441,6 +463,11 @@ class FeatureToggle extends QuickSettings.QuickToggle {
 
         this.fim = fim;
         
+	this.menu.setHeader('selection-mode-symbolic', _('DownFall'));
+	
+	this.menu.addMenuItem(new PopupMenu.PopupSeparatorMenuItem());
+	const settingsItem = this.menu.addAction('Open Settings', () => extensionObject.openPreferences());
+	
         // Binding the toggle to a GSettings key
 	this._settings = extensionObject.getSettings();